fix: report more detailed errors when the command fails
When the formatting command fails, we would like to see its output (when
running formatting automatically, not as a part of an autocmd), so that
we know what’s wrong and can fix it.

This change makes it so that we print the output at the end.

This also takes care of interrupting commands that run for unreasonably
long periods of time. In any case, it is always possible to stop the job
with CTRL-C.